Add a dummy login_required function for development
I haven't thought of how to structure deployments yet.  As I'm planning
to import this under a sub-path in an existing site, it makes the most
sense to use a blueprint.  But I also want this to be able to run as a
standalone app.

Part of the issue is that I need to pass in a login_required decorator
from outside.  See the referenced ticket.

References: https://todo.sr.ht/~runxiyu/fbfp/1